requirements.txt
random
google-generativeai
python-dotenv
discord.py
responses
flask